recuperation projet pyriam
recuperation des donner du projet complet

+++ src/main/java/fr/dawan/projetcompagnieaerienne/entities/Aeroport.java
@@ -1,37 +1,41 @@
 package fr.dawan.projetcompagnieaerienne.entities;
 
 import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.Version;
 
 @Entity
-@Table(name = "aeroports")
-public class Aeroport implements Serializable {
+@Table(name="aeroports")
+public class Aeroport implements Serializable{
 
     private static final long serialVersionUID = 1L;
-
+    
     @Version
     private int version;
-
+    
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id_aeroport")
-    private int idAeroport;
-
-    @Column(name = "code_IATA", nullable = false, length = 3)
+    @GeneratedValue(strategy=GenerationType.IDENTITY)
+    @Column(name="id_aeroport")
+    private long idAeroport;
+    
+    @Column(name="code_IATA", nullable=false, length=3)
     private String codeIATA;
-
+    
     private String nom;
-
+    
     private String pays;
 
-    // constructeurs
+    
+    //constructeurs
     public Aeroport(String nom, String codeIATA, String pays) {
         this.nom = nom;
         this.codeIATA = codeIATA;
@@ -39,15 +43,16 @@ public class Aeroport implements Serializable {
     }
 
     public Aeroport() {
-
+        
     }
 
-    // getters & setters
-    public int getIdAeroport() {
+    
+    //getters & setters
+    public long getIdAeroport() {
         return idAeroport;
     }
 
-    public void setIdAeroport(int idAeroport) {
+    public void setIdAeroport(long idAeroport) {
         this.idAeroport = idAeroport;
     }
 
@@ -74,5 +79,8 @@ public class Aeroport implements Serializable {
     public void setPays(String pays) {
         this.pays = pays;
     }
+    
+    
+    
 
 }
